Revert DTO renaming
