+++ src/main/java/hr/fer/rsikspr/chatbot/model/entity/Message.java
@@ -1,4 +1,4 @@
-package hr.fer.rsikspr.chatbot.model;
+package hr.fer.rsikspr.chatbot.model.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
